+++ ginum-backend-main/src/main/java/com/example/GinumApps/repository/SupplierRepository.java
@@ -3,10 +3,11 @@ package com.example.GinumApps.repository;
 import com.example.GinumApps.model.Supplier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;import java.util.Optional;
+import java.util.List;
+import java.util.Optional;
 
-public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
+public interface SupplierRepository extends JpaRepository<Supplier, Long> {
     List<Supplier> findByCompany_CompanyId(Integer companyId);
 
-//    Optional<Supplier> findByIdAndCompany_Id(Long id, Integer companyId);
+    // Optional<Supplier> findByIdAndCompany_Id(Long id, Integer companyId);
 }
